+++ common/source/servers/Proxy.cpp
@@ -123,8 +123,16 @@ void Proxy::run() {
       // public
       if (resp_from_client_socket.error == false and
           resp_from_client_socket.time_out == false) {
-        m_communicators[SocketRole::SERVER]->send(
-            *resp_from_client_socket.message);
+        if (not resp_from_client_socket.message) {
+          DL_ERROR(
+              m_log_context,
+              "Proxy::run - Something is wrong, message "
+                  << "response is not defined but no timeouts or errors were "
+                  << "triggered, unable to send to server.");
+        } else {
+          m_communicators[SocketRole::SERVER]->send(
+              *resp_from_client_socket.message);
+        }
       }
 
       // If there are operations that need to happen on incoming messages,
@@ -137,11 +145,19 @@ void Proxy::run() {
       // ... - Serv Sock - Proxy ------ Client Sock - Serv Sock - Inter App
       if (resp_from_server_socket.error == false and
           resp_from_server_socket.time_out == false) {
-        for (auto &in_operator : m_incoming_operators) {
-          in_operator->execute(*resp_from_server_socket.message);
+        if (not resp_from_server_socket.message) {
+          DL_ERROR(
+              m_log_context,
+              "Proxy::run - Something is wrong, message "
+                  << "response is not defined but no timeouts or errors were "
+                  << "triggered, unable to operate and send to client.");
+        } else {
+          for (auto &in_operator : m_incoming_operators) {
+            in_operator->execute(*resp_from_server_socket.message);
+          }
+          m_communicators[SocketRole::CLIENT]->send(
+              *resp_from_server_socket.message);
         }
-        m_communicators[SocketRole::CLIENT]->send(
-            *resp_from_server_socket.message);
       }
 
     } catch (TraceException &e) {