+++ compose/build_repo_images_for_compose.sh
@@ -7,10 +7,10 @@ SOURCE=$(dirname "$SCRIPT")
 PROJECT_ROOT=$(realpath "${SOURCE}/../")
 . "${PROJECT_ROOT}/scripts/dependency_versions.sh"
 
-cd ${PROJECT_ROOT}/external/globus-connect-server-deploy/docker
-git checkout "DATAFED_GCS_SUBMODULE_VERSION"
+cd "${PROJECT_ROOT}/external/globus-connect-server-deploy/docker"
+git checkout "$DATAFED_GCS_SUBMODULE_VERSION"
 docker build --progress plain --tag "gcs-ubuntu-base:latest" - < "./docker-files/Dockerfile.ubuntu-20.04"
-cd ${PROJECT_ROOT}
+cd "${PROJECT_ROOT}"
 docker build \
   -f "${PROJECT_ROOT}/docker/Dockerfile.dependencies" \
   "${PROJECT_ROOT}" \