+++ CHANGELOG.md
@@ -10,7 +10,8 @@
    smaller images
 4. [912] - Adding working compose instance for core metadata services.
 5. [937] - Working metadata services running together as part of CI
-6. [945] - Added C functions for comparing POSIX path in Authz module
+6. [946] - Added docker compose for DataFed Repository and for Metadata Services
+7. [945] - Added C functions for comparing POSIX path in Authz module
 
 ## PATCH Bug fixes/Technical Debt/Documentation